requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
dateparser==1.2.1
fonttools==4.57.0
frozenlist==1.6.0
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
multidict==6.4.3
numpy==2.0.2
packaging==25.0
pillow==11.2.1
propcache==0.3.1
pycryptodome==3.22.0
pyparsing==3.2.3
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
zipp==3.21.0
